updated files and working uptill login

+++ api/user/customer.py
@@ -2,6 +2,7 @@ from flask import Blueprint, Flask, render_template, request, jsonify
 from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy.sql import text
 from datetime import datetime
+from api.util.validator import validate_input_address_format, validate_integer_text_format
 from db import db
 
 customer = Blueprint('customer', __name__)
@@ -13,6 +14,10 @@ cust = []
 def getCustomerByUserID():
     name = request.args.get('name')
     address = request.args.get('address')
+    if not validate_input_address_format(address):
+        return render_template("error.html", error = {'status': 'validation error', 'message': 'address not in correct format'})
+    if not validate_integer_text_format(name):
+        return render_template("error.html", error = {'status': 'validation error', 'message': 'name not in correct format'})
     # name = request.form['name']
     # address = request.form['address']
     result = db.session.execute(text('''Select * 
@@ -20,10 +25,11 @@ def getCustomerByUserID():
                                      where c.name = :name and c.bill_addr = :address'''), params={'name': name, 'address': address}).fetchone()
     
     # dic = {'name': result[1], 'address': result[2]}
-    curr_cust = {"cust_id":result[0], "name":name, "bill_addr": address}
+    cust_id = result[0] if result else render_template("error.html", error = {'status': 'unexpected error', 'message': 'Unexpected error'})
+    curr_cust = {"cust_id": cust_id, "name":name, "bill_addr": address}
     if curr_cust not in cust:
         cust.append(curr_cust)
-    return render_template('customer.html', name=name,cust_id=result[0])
+    return render_template('customer.html', name=name,cust_id=cust_id)
 
 @customer.route("/user/<cust_id>")
 def customerDetails(cust_id):
@@ -33,7 +39,8 @@ def customerDetails(cust_id):
 # get all locations for a customer
 @customer.route("/user/<cust_id>/locations", methods=['GET'])
 def getLocationsByUserID(cust_id):
-    print("here")
+    if not validate_integer_text_format(cust_id):
+        return render_template("error.html", error = {'status': 'validation error', 'message': 'cust id not in correct format'})
     result = db.session.execute(text('''Select * from service_loc where cust_id = :id'''), params={'id': cust_id}).fetchall()
     results = []
     for r in result:
@@ -53,6 +60,8 @@ def getLocationsByUserID(cust_id):
 
 @customer.route("/user/<cust_id>/locations/<loc_id>/devices", methods=['GET'])
 def getDevicesByLoc(cust_id, loc_id):
+    if not validate_integer_text_format(cust_id) or not validate_integer_text_format(loc_id):
+        return render_template("error.html", error = {'status': 'validation error', 'message': 'some id not in correct format'})
     result = db.session.execute(text('''Select d.dev_id, m.model_name, m.model_type 
                                      from device d 
                                      natural join service_loc sl 
@@ -69,21 +78,21 @@ def getDevicesByLoc(cust_id, loc_id):
     return results
 
 # get devices list with user id
-@customer.route("/user/<cust_id>/devices", methods=['GET'])
-def getDevicesByUserID(cust_id):
-    result = db.session.execute(text('''Select d.dev_id, m.model_name, m.model_type 
-                                     from device d 
-                                     natural join service_loc sl 
-                                     natural join model m
-                                     where sl.cust_id = :id'''), params={'id': cust_id}).fetchall()
-    # results = []
-    # for r in result:
-    #     dic = {}
-    #     dic['dev_id'] = r[0]
-    #     dic['model_name'] = r[1]
-    #     dic['model_type'] = r[2]
-    return result
-    # return render_template('model.html')
+# @customer.route("/user/<cust_id>/devices", methods=['GET'])
+# def getDevicesByUserID(cust_id):
+#     result = db.session.execute(text('''Select d.dev_id, m.model_name, m.model_type 
+#                                      from device d 
+#                                      natural join service_loc sl 
+#                                      natural join model m
+#                                      where sl.cust_id = :id'''), params={'id': cust_id}).fetchall()
+#     # results = []
+#     # for r in result:
+#     #     dic = {}
+#     #     dic['dev_id'] = r[0]
+#     #     dic['model_name'] = r[1]
+#     #     dic['model_type'] = r[2]
+#     return result
+#     # return render_template('model.html')
     
 
 # add user
@@ -108,6 +117,8 @@ def addLocation():
 # delete user
 @customer.route("/user/<cust_id>/del", methods=['DELETE'])
 def deleteDevice(cust_id):
+    if not validate_integer_text_format(cust_id):
+        return render_template("error.html", error = {'status': 'validation error', 'message': 'cust id not in correct format'})
     try:
         db.session.execute(text("Delete from customer where cust_id=:id"), params={'id': cust_id})
         db.session.commit()
@@ -118,5 +129,6 @@ def deleteDevice(cust_id):
     
 def getNextCustId():
     # write a sequence in the database and get id from it
-    last_db_id = db.session.execute(text("Select cust_id from customer order by cust_id desc")).fetchone()[0]
+    last_db_id = db.session.execute(text("Select cust_id from customer order by cust_id desc")).fetchone()
+    last_db_id = last_db_id[0] if last_db_id else render_template("error.html", error = {'status': 'unexpected error', 'message': 'Unexpected error'})
     return str(int(last_db_id)+1)
